feat(runner): add subcommands
